[BlockChain] Start sync system

+++ blockchain/src/main/kotlin/com/supbank/blockchain/repos/TransactionRepository.kt
@@ -3,7 +3,7 @@ package com.supbank.blockchain.repos
 import com.supbank.blockchain.models.Transaction
 import org.springframework.data.repository.CrudRepository
 
-interface TransactionRepository: CrudRepository<Transaction, Int> {
+interface TransactionRepository: CrudRepository<Transaction, Long> {
 
     fun findAllByMinedFalse() : List<Transaction>
